Extract most of ScriptAPI from SnippetsBar
Instead of working through a chain of React props, now talks directly to the editor. And the API itself lives in its own file, to be converted to a React Class later (with the dangling snippetUploadFile, which is still to be extracted).
